Extract optimize_template_variables + rating-slot offset math to modules/output_optimize.py
FIRST ELEPHANT DOWN.  Eleventh refactor pass at modules/output.py --
this one carves out the ~400-line optimize_template_variables
function, the pre-YAML pruning pass that strips template variables
whose values already match their schema defaults.

Before: one 404-line def with 4 nested closures, one of which
(_ensure_explicit_ratings_offsets) is itself a 200-line beast holding
six more nested r{slot}{axis} closures for rating overlay offset math.
Total: 10 levels of function-inside-function-inside-function.  Not
even Guido's grandmother could love it.

After: modules/output_optimize.py with:

Small pure helpers (hoisted to module-level, independently testable):
  * _to_offset_number             -- coerce to int/float w/ fallback
  * _is_ratings_entry             -- 'ratings'/'overlay_ratings*' matcher
  * _normalize_choice             -- lowercase-strip w/ fallback
  * _reorder_ratings_template_vars -- canonical key order for stable YAML

Constants hoisted out of function-local scope:
  * _RATING_CONSTANTS             -- the 10-key offset lookup table
  * _RATINGS_PREFERRED_ORDER      -- the 45-key preferred-order tuple
  * _RATINGS_ALWAYS_KEEP_KEYS     -- frozenset for O(1) membership

Offset math walker:
  * _ensure_explicit_ratings_offsets -- keeps the six r{slot}{axis}
    nested closures because they close over 7 layout-state variables;
    dataclass-ification is a follow-up.  The math is a straight lift
    from the original -- zero behavior change.

Orchestrator broken into tiny SRP-clean pieces:
  * _resolve_libraries_dict       -- unwrap nested-or-flat libraries shape
  * _pick_collection_defaults     -- movie/show fallback to 'all'
  * _pick_overlay_defaults        -- per-level merging with offsets
  * _apply_always_keep            -- force ratings + builder_level keys
  * _optimize_library_level       -- prune library-scope tv
  * _optimize_collection_files    -- prune per-collection-file tv
  * _optimize_overlay_files       -- prune per-overlay-file tv (ratings-aware)
  * optimize_template_variables   -- 20-line orchestrator that just loops

Impact:
- modules/output.py: 2768 -> 2363 (-405 / -14.6%)
- modules/output_optimize.py: ~557 lines new
  (docstrings, decomposition, and hoisted constants added clarity
  but ate lines.  Trade-off is worth it -- SRP-split code is testable
  per-piece.)

Cumulative on output.py: 3833 -> 2363 (-1470 / -38.4%)

All 1077 tests pass. Ruff + black clean.

Zero API break: output.py still re-imports optimize_template_variables
as before.  Zero external callers today (only build_config in
output.py uses it), so no compatibility ceremony needed.
