+++ modules/output.py
@@ -46,6 +46,7 @@ from modules.output_headers import (  # noqa: F401 -- re-exported so output.<nam
     add_border_to_ascii_art,
     section_heading,
 )
+from modules.output_optimize import optimize_template_variables
 from modules.output_playlists import (  # noqa: F401 -- re-exported so output.<name> keeps working
     PLAYLIST_KEYED_TEMPLATE_VAR_SPECS,
     PLAYLIST_SHARED_TEMPLATE_VAR_SPECS,
@@ -121,412 +122,6 @@ LIBRARY_SONARR_FIELDS = {
     "sonarr_path": "string",
     "plex_path": "string",
 }
-
-
-def optimize_template_variables(config_data, library_types=None):
-    def _to_offset_number(value, fallback):
-        if isinstance(value, bool):
-            return fallback
-        if isinstance(value, (int, float)):
-            return value
-        if isinstance(value, str):
-            stripped = value.strip()
-            if not stripped:
-                return fallback
-            try:
-                return int(stripped)
-            except ValueError:
-                try:
-                    return float(stripped)
-                except ValueError:
-                    return fallback
-        return fallback
-
-    def _is_ratings_entry(default_name):
-        return isinstance(default_name, str) and (default_name == "ratings" or default_name.startswith("overlay_ratings"))
-
-    def _ensure_explicit_ratings_offsets(tv, defaults):
-        if not isinstance(tv, dict):
-            return
-        slot_ids = []
-        for idx in ("1", "2", "3"):
-            rating_key = f"rating{idx}"
-            image_key = f"{rating_key}_image"
-            if rating_key in tv and image_key in tv:
-                slot_ids.append(idx)
-        if not slot_ids:
-            return
-        defaults = defaults or {}
-
-        def _normalize_choice(value, fallback):
-            raw = str(value if value is not None else fallback).strip().lower()
-            return raw
-
-        def _slot_enabled(slot):
-            rating_val = _normalize_choice(tv.get(f"rating{slot}"), "")
-            image_val = _normalize_choice(tv.get(f"rating{slot}_image"), "")
-            if rating_val in ("", "none") or image_val in ("", "none"):
-                return False
-            return True
-
-        alignment = _normalize_choice(tv.get("rating_alignment", defaults.get("rating_alignment", "vertical")), "vertical")
-        if alignment not in ("horizontal", "vertical"):
-            alignment = "vertical"
-        h_pos = _normalize_choice(tv.get("horizontal_position", defaults.get("horizontal_position", "left")), "left")
-        if h_pos not in ("left", "center", "right"):
-            h_pos = "left"
-        v_pos = _normalize_choice(tv.get("vertical_position", defaults.get("vertical_position", "center")), "center")
-        if v_pos not in ("top", "center", "bottom"):
-            v_pos = "center"
-
-        if isinstance(defaults, dict):
-            defaults["back_width"] = 270 if alignment == "horizontal" else 160
-            defaults["back_height"] = 80 if alignment == "horizontal" else 160
-            defaults["addon_position"] = "left" if alignment == "horizontal" else "top"
-            defaults["horizontal_offset"] = 0 if h_pos == "center" else 15
-            defaults["vertical_offset"] = 0 if v_pos == "center" else 15
-
-        rating_constants = {
-            "standard": 30,
-            "center": 0,
-            "v2": 235,
-            "v3": 440,
-            "cv2": 105,
-            "cv3": 205,
-            "h2": 345,
-            "h3": 660,
-            "ch2": 160,
-            "ch3": 335,
-        }
-
-        none1 = not _slot_enabled("1")
-        none2 = not _slot_enabled("2")
-        none3 = not _slot_enabled("3")
-
-        def r1h():
-            if alignment == "vertical" and h_pos == "center":
-                return rating_constants["center"]
-            if alignment == "horizontal" and h_pos == "center" and none2 and none3:
-                return rating_constants["center"]
-            if alignment == "horizontal" and h_pos == "center" and none2:
-                return -rating_constants["ch2"]
-            if alignment == "horizontal" and h_pos == "center" and none3:
-                return -rating_constants["ch2"]
-            if alignment == "horizontal" and h_pos == "center":
-                return -rating_constants["ch3"]
-            if alignment == "horizontal" and h_pos == "right" and none2 and none3:
-                return rating_constants["standard"]
-            if alignment == "horizontal" and h_pos == "right" and none2:
-                return rating_constants["h2"]
-            if alignment == "horizontal" and h_pos == "right" and none3:
-                return rating_constants["h2"]
-            if alignment == "horizontal" and h_pos == "right":
-                return rating_constants["h3"]
-            return rating_constants["standard"]
-
-        def r1v():
-            if alignment == "horizontal" and v_pos == "center":
-                return rating_constants["center"]
-            if alignment == "vertical" and v_pos == "center" and none2 and none3:
-                return rating_constants["center"]
-            if alignment == "vertical" and v_pos == "center" and none2:
-                return -rating_constants["cv2"]
-            if alignment == "vertical" and v_pos == "center" and none3:
-                return -rating_constants["cv2"]
-            if alignment == "vertical" and v_pos == "center":
-                return -rating_constants["cv3"]
-            if alignment == "vertical" and v_pos == "bottom" and none2 and none3:
-                return rating_constants["standard"]
-            if alignment == "vertical" and v_pos == "bottom" and none2:
-                return rating_constants["v2"]
-            if alignment == "vertical" and v_pos == "bottom" and none3:
-                return rating_constants["v2"]
-            if alignment == "vertical" and v_pos == "bottom":
-                return rating_constants["v3"]
-            return rating_constants["standard"]
-
-        def r2h():
-            if alignment == "vertical" and h_pos == "center":
-                return rating_constants["center"]
-            if alignment == "horizontal" and h_pos == "center" and none1 and none3:
-                return rating_constants["center"]
-            if alignment == "horizontal" and h_pos == "center" and none1:
-                return -rating_constants["ch2"]
-            if alignment == "horizontal" and h_pos == "center" and none3:
-                return rating_constants["ch2"]
-            if alignment == "horizontal" and h_pos == "center":
-                return rating_constants["center"]
-            if alignment == "horizontal" and h_pos == "right" and none1 and none3:
-                return rating_constants["standard"]
-            if alignment == "horizontal" and h_pos == "right" and none3:
-                return rating_constants["standard"]
-            if alignment == "horizontal" and h_pos == "right":
-                return rating_constants["h2"]
-            if alignment == "horizontal" and h_pos == "left" and none1:
-                return rating_constants["standard"]
-            if alignment == "horizontal" and h_pos == "left":
-                return rating_constants["h2"]
-            return rating_constants["standard"]
-
-        def r2v():
-            if alignment == "horizontal" and v_pos == "center":
-                return rating_constants["center"]
-            if alignment == "vertical" and v_pos == "center" and none1 and none3:
-                return rating_constants["center"]
-            if alignment == "vertical" and v_pos == "center" and none1:
-                return -rating_constants["cv2"]
-            if alignment == "vertical" and v_pos == "center" and none3:
-                return rating_constants["cv2"]
-            if alignment == "vertical" and v_pos == "center":
-                return rating_constants["center"]
-            if alignment == "vertical" and v_pos == "bottom" and none1 and none3:
-                return rating_constants["standard"]
-            if alignment == "vertical" and v_pos == "bottom" and none1:
-                return rating_constants["v2"]
-            if alignment == "vertical" and v_pos == "bottom" and none3:
-                return rating_constants["standard"]
-            if alignment == "vertical" and v_pos == "bottom":
-                return rating_constants["v2"]
-            if alignment == "vertical" and v_pos == "top" and none1:
-                return rating_constants["standard"]
-            if alignment == "vertical" and v_pos == "top":
-                return rating_constants["v2"]
-            return rating_constants["standard"]
-
-        def r3h():
-            if alignment == "vertical" and h_pos == "center":
-                return rating_constants["center"]
-            if alignment == "horizontal" and h_pos == "center" and none1 and none2:
-                return rating_constants["center"]
-            if alignment == "horizontal" and h_pos == "center" and none1:
-                return rating_constants["ch2"]
-            if alignment == "horizontal" and h_pos == "center" and none2:
-                return rating_constants["ch2"]
-            if alignment == "horizontal" and h_pos == "center":
-                return rating_constants["ch3"]
-            if alignment == "horizontal" and h_pos == "left" and none1 and none2:
-                return rating_constants["standard"]
-            if alignment == "horizontal" and h_pos == "left" and none1:
-                return rating_constants["h2"]
-            if alignment == "horizontal" and h_pos == "left" and none2:
-                return rating_constants["h2"]
-            if alignment == "horizontal" and h_pos == "left":
-                return rating_constants["h3"]
-            return rating_constants["standard"]
-
-        def r3v():
-            if alignment == "horizontal" and v_pos == "center":
-                return rating_constants["center"]
-            if alignment == "vertical" and v_pos == "center" and none1 and none2:
-                return rating_constants["center"]
-            if alignment == "vertical" and v_pos == "center" and none1:
-                return rating_constants["cv2"]
-            if alignment == "vertical" and v_pos == "center" and none2:
-                return rating_constants["cv2"]
-            if alignment == "vertical" and v_pos == "center":
-                return rating_constants["cv3"]
-            if alignment == "vertical" and v_pos == "top" and none1 and none2:
-                return rating_constants["standard"]
-            if alignment == "vertical" and v_pos == "top" and none1:
-                return rating_constants["v2"]
-            if alignment == "vertical" and v_pos == "top" and none2:
-                return rating_constants["v2"]
-            if alignment == "vertical" and v_pos == "top":
-                return rating_constants["v3"]
-            return rating_constants["standard"]
-
-        computed = {
-            "1": {"h": r1h(), "v": r1v()},
-            "2": {"h": r2h(), "v": r2v()},
-            "3": {"h": r3h(), "v": r3v()},
-        }
-
-        for idx in slot_ids:
-            h_key = f"rating{idx}_horizontal_offset"
-            v_key = f"rating{idx}_vertical_offset"
-            if isinstance(defaults, dict):
-                defaults[h_key] = int(round(computed[idx]["h"]))
-                defaults[v_key] = int(round(computed[idx]["v"]))
-            if h_key not in tv:
-                tv[h_key] = int(round(computed[idx]["h"]))
-            if v_key not in tv:
-                tv[v_key] = int(round(computed[idx]["v"]))
-
-    def _reorder_ratings_template_vars(entry):
-        if not isinstance(entry, dict):
-            return
-        default_name = entry.get("default", "")
-        if not _is_ratings_entry(default_name):
-            return
-        tv = entry.get("template_variables")
-        if not isinstance(tv, dict) or not tv:
-            return
-        preferred_order = [
-            "builder_level",
-            "rating1",
-            "rating1_image",
-            "rating1_font",
-            "rating1_font_size",
-            "rating1_font_color",
-            "rating1_stroke_width",
-            "rating1_stroke_color",
-            "rating1_horizontal_offset",
-            "rating1_vertical_offset",
-            "rating2",
-            "rating2_image",
-            "rating2_font",
-            "rating2_font_size",
-            "rating2_font_color",
-            "rating2_stroke_width",
-            "rating2_stroke_color",
-            "rating2_horizontal_offset",
-            "rating2_vertical_offset",
-            "rating3",
-            "rating3_image",
-            "rating3_font",
-            "rating3_font_size",
-            "rating3_font_color",
-            "rating3_stroke_width",
-            "rating3_stroke_color",
-            "rating3_horizontal_offset",
-            "rating3_vertical_offset",
-            "horizontal_position",
-            "vertical_position",
-            "horizontal_offset",
-            "vertical_offset",
-            "flag_alignment",
-            "back_align",
-            "back_color",
-            "back_height",
-            "back_width",
-            "back_line_color",
-            "back_line_width",
-            "back_padding",
-            "back_radius",
-            "use_subtitles",
-        ]
-        ordered = {}
-        for key in preferred_order:
-            if key in tv:
-                ordered[key] = tv[key]
-        for key in tv:
-            if key not in ordered:
-                ordered[key] = tv[key]
-        entry["template_variables"] = ordered
-
-    libraries_section = config_data.get("libraries", {})
-    if isinstance(libraries_section, dict) and isinstance(libraries_section.get("libraries"), dict):
-        libraries = libraries_section.get("libraries")
-    elif isinstance(libraries_section, dict):
-        libraries = libraries_section
-    else:
-        libraries = None
-    if not isinstance(libraries, dict):
-        return config_data
-
-    collection_defaults = _build_collection_defaults()
-    overlay_defaults = _build_overlay_defaults()
-    attribute_defaults = _build_attribute_defaults()
-
-    for library_name, library_data in libraries.items():
-        if not isinstance(library_data, dict):
-            continue
-
-        library_type = None
-        if isinstance(library_types, dict):
-            library_type = library_types.get(library_name)
-
-        tv = library_data.get("template_variables")
-        if isinstance(tv, dict):
-            pruned = _prune_template_variables(tv, attribute_defaults)
-            if pruned:
-                library_data["template_variables"] = pruned
-            else:
-                library_data.pop("template_variables", None)
-
-        collection_files = library_data.get("collection_files")
-        if isinstance(collection_files, list):
-            for entry in collection_files:
-                if not isinstance(entry, dict):
-                    continue
-                tv = entry.get("template_variables")
-                if not isinstance(tv, dict):
-                    continue
-                defaults = None
-                if isinstance(collection_defaults, dict):
-                    if library_type in ("movie", "show"):
-                        defaults = collection_defaults.get(library_type, {}).get(entry.get("default"))
-                    if defaults is None:
-                        defaults = collection_defaults.get("all", {}).get(entry.get("default"))
-                if not defaults:
-                    continue
-                pruned = _prune_template_variables(tv, defaults)
-                if entry.get("default") == "year" and "data_ending" in tv:
-                    pruned["data_ending"] = tv.get("data_ending")
-                if pruned:
-                    entry["template_variables"] = pruned
-                else:
-                    entry.pop("template_variables", None)
-
-        overlay_files = library_data.get("overlay_files")
-        if isinstance(overlay_files, list):
-            for entry in overlay_files:
-                if not isinstance(entry, dict):
-                    continue
-                tv = entry.get("template_variables")
-                if not isinstance(tv, dict):
-                    continue
-                defaults_entry = overlay_defaults.get(entry.get("default"))
-                if not defaults_entry:
-                    continue
-                defaults = dict(defaults_entry.get("defaults", {}))
-
-                overlay_level = None
-                if isinstance(tv.get("builder_level"), str) and tv.get("builder_level"):
-                    overlay_level = tv.get("builder_level")
-                elif library_type:
-                    overlay_level = "movie" if library_type == "movie" else "show"
-
-                if overlay_level:
-                    type_defaults = defaults_entry.get("defaults_by_type", {}).get(overlay_level)
-                    if type_defaults:
-                        defaults = dict(type_defaults)
-                    offsets = defaults_entry.get("offsets_by_type", {}).get(overlay_level)
-                    if offsets:
-                        defaults.update(offsets)
-
-                if _is_ratings_entry(entry.get("default")):
-                    _ensure_explicit_ratings_offsets(tv, defaults)
-
-                pruned = _prune_template_variables(tv, defaults)
-                always_keep = set()
-                if _is_ratings_entry(entry.get("default")):
-                    always_keep.update(
-                        {
-                            "builder_level",
-                            "rating1",
-                            "rating1_image",
-                            "rating2",
-                            "rating2_image",
-                            "rating3",
-                            "rating3_image",
-                            "horizontal_position",
-                            "vertical_position",
-                        }
-                    )
-                if "builder_level" in tv:
-                    always_keep.add("builder_level")
-                if always_keep:
-                    for key in always_keep:
-                        if key in tv:
-                            pruned[key] = tv[key]
-                if pruned:
-                    entry["template_variables"] = pruned
-                    _reorder_ratings_template_vars(entry)
-                else:
-                    entry.pop("template_variables", None)
-
-    return config_data
 
 
 def build_libraries_section(
